Created directories so the files in download_files are separated. Tests sorted and added one more 'test_read_file_is_not_present'
