+++ README.md
@@ -61,4 +61,4 @@ We assume you've got pip installed and python 3.8 or higher, the SO used is ubun
         - to run the tests: python -m unittest tests.test_download_files
 
 Finally you can just run the command: 
-    - python download_files/download_files.py
+    - python -m download_files.download_files
